Fastlane/GH build improvements: Add sync action and keepalive action, align with Loop dev improvements (#46)
* Add update.yml. Runs nightly to check for new commits.

If new commits are found, the target branch on forks is synced with upstream, and build_iAPS.yml is launched. Updates will hence be made available in TestFlight within less than a day. Users will have to decide if they want to enable Automatic Updates in the  TestFlight app or not.

Scheduled workflows in GitHub actions will be cancelled in the event of 60 days of no repository activity. A "keep alive" action will generate empty commits on upstream if there is no activity after 50 days to avoid inactivation of the scheduled workflow. This commit will be picked up by update.yml on forks the next day, so that there is repository activity also on forks.

The reason for running the keepalive action on upstream only, is that upstream and forks should then be on the same commit ref. This would not always be the case if the keepalive action ran on the forks instead. But then the scheduled workflow must be enabled in the upstream repo (Artificial-Pancreas/iAPS).

Changes to build_iAPS.yml:
- Added "on: workflow_call" to allow it to be launched by update.yml
- Removed the "secrets" job", this can still be run from the identifiers and certificates workflows to validate repository secrets.
- Removed checkout option "submodules: recursive"

* Schedule build of iAPS at 04:30 UTC on the 1st  every month

The workflow will not run on schedule on forks until the workflow is activated. It can later easily be inactivated/re-activated without code changes to cancel scheduled runs by selecting "Disable workflow" / "Enable workflow" in Actions.

* build_iAPS: add missing env: TARGET_BRANCH

* update.yml and build_iAPS.yml: Use current branch name as TARGET_BRANCH and UPSTREAM_BRANCH

* Run the keepalive action on all repositories

This helps forks not to rely on the upstream repo to keep scheduled workflows running

* validate_secrets.yml: Migrate validation improvements from Loop

* create_certs.yml: Add branch name to workflow name, change job name

* add_identifiers.yml: Add branch name to workflow name, change job name

* build_iAPS.yml: Add validation as required job, migrate @bjornoleh build_IAPS.yml

* update.yml: Add validation as required job, migrate @bjornoleh update.yml

* Fastfile: Add matching to secrets validation lane to verify MATCH_PASSWORD

* Add update.yml. Runs nightly to check for new commits.

If new commits are found, the target branch on forks is synced with upstream, and build_iAPS.yml is launched. Updates will hence be made available in TestFlight within less than a day. Users will have to decide if they want to enable Automatic Updates in the  TestFlight app or not.

Scheduled workflows in GitHub actions will be cancelled in the event of 60 days of no repository activity. A "keep alive" action will generate empty commits on upstream if there is no activity after 50 days to avoid inactivation of the scheduled workflow. This commit will be picked up by update.yml on forks the next day, so that there is repository activity also on forks.

The reason for running the keepalive action on upstream only, is that upstream and forks should then be on the same commit ref. This would not always be the case if the keepalive action ran on the forks instead. But then the scheduled workflow must be enabled in the upstream repo (Artificial-Pancreas/iAPS).

Changes to build_iAPS.yml:
- Added "on: workflow_call" to allow it to be launched by update.yml
- Removed the "secrets" job", this can still be run from the identifiers and certificates workflows to validate repository secrets.
- Removed checkout option "submodules: recursive"

* Schedule build of iAPS at 04:30 UTC on the 1st  every month

The workflow will not run on schedule on forks until the workflow is activated. It can later easily be inactivated/re-activated without code changes to cancel scheduled runs by selecting "Disable workflow" / "Enable workflow" in Actions.

* build_iAPS: add missing env: TARGET_BRANCH

* update.yml and build_iAPS.yml: Use current branch name as TARGET_BRANCH and UPSTREAM_BRANCH

* Run the keepalive action on all repositories

This helps forks not to rely on the upstream repo to keep scheduled workflows running

* Harmonise workflows with LoopKit/LoopWorkspace

* Update Fastlane to 2.215.0

Among other improvements, this should fix the WWDR issue.

Commands used to install bundler locally and update dependencies:

sudo gem pristine ffi
sudo gem install bundler
sudo bundle install
sudo bundle update fastlane

* Automatically sync fork before building if vars.SCHEDULED_SYNC == 'true'

Automated sync of fork before building requires setting a SCHEDULED_SYNC repository variable to 'true'

Build job:
change condition from vars.SCHEDULED_SYNC != 'false' to vars.SCHEDULED_SYNC == 'true'

for the build job and following steps:

- name: Checkout Repo for syncing
- name: Sync upstream changes
- name: New commits found
- name: No new commits
- name: Show value of 'has_new_commits'

---------

Co-authored-by: Deniz Cengiz <48965855+dnzxy@users.noreply.github.com>